requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.28.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
openpyxl>=3.1.0
sentence-transformers>=2.0.0
numpy>=1.24.0
lxml>=4.9.0
